requirements.txt
telethon==1.34.0
yt-dlp==2025.8.11
python-dotenv==1.0.0
gallery-dl==1.30.2

# System requirements (install separately):
# - ffmpeg (for high-quality video processing and TikTok slideshow creation)
# Ubuntu/Debian: sudo apt install ffmpeg
# macOS: brew install ffmpeg
# Windows: Download from https://ffmpeg.org/download.html

# gallery-dl is used for downloading TikTok photo slideshows with all images and audio